requirements.txt
fastai
torch
numpy
Pillow
